+++ bin/main.dart
@@ -3,8 +3,8 @@ import 'package:palace/palace.dart';
 Future<void> main(List<String> args) async {
   final palace = Palace();
   palace.get('/', (req, res) => 'Long Live The Queen 👑');
-  palace.get('/login', (req, res) {
-    return NotFound();
+  palace.post('/login', (req, res) {
+    return req.body;
   });
   return palace.openGates();
 }