+++ lib/config.dart
@@ -1,7 +1,10 @@
 import 'package:palace/palace.dart';
+import 'package:queen_env/queen_env.dart';
 
 abstract class PalaceConfig {
   /// will be called in case of no match with any of the registered endpoints
   static HandlerFunc notFoundHandler = (req, res) => throw NotFound();
-  static bool enaleLogs = true;
+
+  ///
+  static bool get enaleLogs => env('enableLogs') == 'true';
 }