+++ lib/src/http/exten/req.dart
@@ -1,0 +1,5 @@
+import 'package:palace/src/http/request.dart';
+
+extension ReqGetters on Request {
+  String get contentType => request.headers.contentType.toString();
+}