+++ lib/src/http/responses/ok.dart
@@ -1,7 +1,7 @@
 import 'package:palace/src/imp/palace_response.dart';
 
-class Created extends PalaceResponse {
-  Created([Object? data])
+class Ok extends PalaceResponse {
+  Ok([Object? data])
       : super(
           statusCode: HttpStatus.ok,
           data: data,